requirements.txt
pandas
Pillow
pytesseract
